refactor: change cmd -> format
This allows the code to align with the documentation.

resolves: #20
