--- cmd/number_man.toml
@@ -1,12 +1,0 @@
-name = "number"
-synopsis = "human number [-g] input"
-description = """
-Breakdown a large number by either grouping it into groups of thousands or
-translating it into english.
-"""
-
-[flags]
-
-  [flags.g]
-  mnemonic = "grouping"
-  desc = "Breaks up large numbers into groupings of thousands"
