+++ main.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	"github.com/andres-lowrie/human/cmd"
+	"github.com/andres-lowrie/human/format"
 )
 
 func main() {
@@ -16,7 +17,7 @@ func main() {
 	// specific similar to `dig`, where `dig` with no args gives all the
 	// information it has, and then something like `dig +short` gives you a whole
 	// lot less
-	handlers := map[string]cmd.Command{"number": cmd.NewNumber(), "size": cmd.NewSize()}
+	handlers := map[string]format.Format{"number": format.NewNumber(), "size": format.NewSize()}
 
 	// Figure out direction and which format
 	// we'll default to the `--from` direction since it might be the most common
